+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * [Breaking]: Introduced the `GoogleApiConfig` class to manage the Google Places API Configuration.
 * [Breaking]: Renamed `onPlaceDetailsWithCoordinatesReceived` to `onPredictionWithCoordinatesReceived`.
-* [Feat]: Add `locationBias` and `locationRestriction` parameters to the query (see more info [in the Google docs](https://developers.google.com/maps/documentation/places/web-service/place-autocomplete#location-bias-restriction))
+* [Feat]: Add `locationBias` and `locationRestriction` parameters to the api config(see more info [in the Google docs](https://developers.google.com/maps/documentation/places/web-service/place-autocomplete#location-bias-restriction))
+* [Feat]: Add `includedPrimaryTypes` parameter to the api config (see more info [in the Google docs](https://developers.google.com/maps/documentation/places/web-service/place-autocomplete#includedPrimaryTypes))
 
 ## 1.1.0